### 2021-08-07 46987ca
Update paper.md
Removed blank orcid entries
