### 2021-08-18 #3588 @crvernon
:wave: @rmorgan10 @crhea93 - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
