### 2021-08-20 #3635 @whedon
[REVIEW]: AstronomicAL: an interactive dashboard for visualisation, integration and classification of data with Active Learning
**Submitting author:** @grant-m-s (<a href="http://orcid.org/0000-0002-8885-4443">Grant Stevens</a>)
**Repository:** <a href="https://github.com/grant-m-s/astronomicAL" target ="_blank">https://github.com/grant-m-s/astronomicAL</a>
**Version:** v1.0
