### 2021-08-20 #3635 @grant-m-s
Thanks @crvernon for setting this up and thanks @crhea93 and @rmorgan10 for agreeing to review AstronomicAL. If either of you have any questions I'll be happy to answer them.
