### 2021-08-24 grant-m-s/AstronomicAL#9 @rmorgan10
[JOSS Review] Paper
This issue is part of the JOSS review going on in https://github.com/openjournals/joss-reviews/issues/3635.

The paper is well written and does an excellent job of introducing astronomicAL. I think two aspects of the paper can be improved.

1. The state of the field discussion seems to be missing. Perhaps some examples where active learning has been applied would be helpful here.

1. The statement of need section seems to list the features of astronomicAL as opposed to demonstrating why it is a needed piece of software. I'm aware that there is significant need for a tool like this, but I think that fact could be communicated better in the paper. For instance, when you add a more descriptive state of the field, that discussion will lend itself nicely to introducing the reason the field needs astronimcAL.
