### 2021-08-24 grant-m-s/AstronomicAL#10 @rmorgan10
[JOSS Review] Installation
This issue is part of the JOSS review going on in https://github.com/openjournals/joss-reviews/issues/3635.

At present, the installation instructions are inaccurate. After cloning, users need to enter the command `cd astronomicAL` such that the environment creation command can access the `requirements.txt` file. Adding this command to the instructions will be sufficient to satisfy the JOSS installation specifications.

That being said, I think the better way to go about installation would be to make your package available on PyPI. The motivation here is to enable users to be able to use astronomicAL in any directory they are working in as opposed to having to move their work to the location of the cloned repository. This change is optional, but could make astronomicAL much more portable.
