### 2021-08-24 grant-m-s/AstronomicAL#11 @rmorgan10
[JOSS Review] General Documentaion
This issue is part of the JOSS review going on in openjournals/joss-reviews#3635.

The methods are sufficiently documented given that users interact with astronomicAL through the interactive browser as opposed to directly with the API. The tutorials are very helpful in this regard.

Just one thing is missing to satisfy JOSS requirements:

1. Community guidelines for how people can contribute, report issues, and seek support. Generally, just a short paragraph in the README will satisfy this requirement.
