### 2021-08-24 #3635 @rmorgan10
Hi All!

@grant-m-s I would like to first congratulate you on an awesome software package! `astronomicAL` is really cool and a great addition to the field as a whole. I've identified 4 small things that will help me to check off all the remaining boxes above:

1. Changes to the paper to clarify the state of the field and the statement of need in grant-m-s/AstronomicAL#9
1. Updates to the installation instructions in grant-m-s/AstronomicAL#10
1. Clarity on contributing and bug reporting in grant-m-s/AstronomicAL#11
1. Verifying the functionalities in the testing suite in grant-m-s/AstronomicAL#12

On all these issues, let's iterate on them in your repo, and when they have been resolved you can close them.

Overall the tutorials are very thorough, the code is sufficiently documented, and I found `astronomicAL` easy to use. Great work!
