### 2021-08-24 e508ff8
Update README to fix Issue #10.
Fix for #10  .
Updated installation instructions.
