### 2021-08-24 587b187
Update README to fix Issue #10.
Fix for #10  .
Updated installation instructions.
