### 2021-08-24 1cb507f
Update README to fix Issue #10.
Fix for #10  .
Updated installation instructions.
