### 2021-08-24 grant-m-s/AstronomicAL#12 @grant-m-s
> Split the test file into several individual files such that each sub-module has its own test file with just a few tests.

I will absolutely do that.

> When I ran the tests, 32 failed.

Whenever I run `pytest` I run it within the repo's main directory (so one dir above `tests/`). This is how I have done it throughout AstronomicAL's development (and what the Travis-CI is set up to do on pushing too). I just ran `pytest` within `tests/`and got a similar output to you so I'm assuming that is what you did.

Could you rerun the tests from within the top level directory? This should then give you all the passing tests.

Thanks again!
