### 2021-08-25 7fe0260
Updated README by adding Contributions section
Fix for #11  .
Added sections for bugs, extensions and feedback.
