### 2021-08-25 b52acd4
Updated README by adding Contributions section
Fix for #11  .
Added sections for bugs, extensions and feedback.
