### 2021-08-25 b037290
Updated README by adding Contributions section
Fix for #11  .
Added sections for bugs, extensions and feedback.
