### 2021-08-25 grant-m-s/AstronomicAL#11 @grant-m-s
An update has now been pushed, let me know if it's sufficient.
