### 2021-08-25 grant-m-s/AstronomicAL#10 @grant-m-s
I made the `cd astronomicAL` change.

> I think the better way to go about installation would be to make your package available on PyPI

This was something I was planning on doing after the JOSS submission.

Let me know if this issue is now closable.
