### 2021-08-25 grant-m-s/AstronomicAL#9 @grant-m-s
Thanks again for the feedback.

This one will take me about a week to get an update through, as I already found it a challenge to get the word count down to what it is currently. Are there any parts of the paper that you feel are unnecessary?
