### 2021-08-25 bee91f0
Update Docs index.rst for issues fixes
Fix for #10 and #11 .
Same changes to docs as in repo README.
