### 2021-08-25 #3635 @crvernon
:mega:  Mid-week rally!  Just checking in to see how things are going.

🙌 Thanks for getting the ball rolling @rmorgan10 !

:open_hands: @crhea93 I am happy to answer any question you may have as well!

Great work on this review so far!
