### 2021-08-25 grant-m-s/AstronomicAL#10 @crhea93
Hey there!
I've followed the new installation instructions
``` git clone https://github.com/grant-m-s/AstronomicAL.git
    cd AstronomicAL
    conda create --name astronomical --file requirements.txt
    conda activate astronomical
```
but I get the following error:
```
remote: Enumerating objects: 1850, done.
remote: Total 1850 (delta 8), reused 0 (delta 0), pack-reused 1833
Resolving deltas: 100% (1088/1088), done.
Collecting package metadata (current_repodata.json): done
Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
Solving environment: failed

PackagesNotFoundError: The following packages are not available from current channels:

  - modal[version='>=0.4.1']
  - datashader==0.12.0

Current channels:

  - https://repo.anaconda.com/pkgs/main/linux-64
  - https://repo.anaconda.com/pkgs/main/noarch
  - https://repo.anaconda.com/pkgs/r/linux-64
  - https://repo.anaconda.com/pkgs/r/noarch

To search for alternate channels that may provide the conda package you're
looking for, navigate to

    https://anaconda.org

and use the search bar at the top of the page.

Could not find conda environment: astronomical
You can list all discoverable environments with `conda info --envs`.
```

The last bit is of course because the environment wasn't properly created, but the conda install failed because it couldn't find those two packages: modal & datashader.
