### 2021-08-25 #3635 @crhea93
@grant-m-s WHAT A PACKAGE! This is extremely impressive work! The documentation is absolutely spot on -- I was able to answer any questions I had about the software and how best to use it by scrolling through the readthedocs pages.

I've added an additional installation issue to the existing GitHub issue created by @rmorgan10 #11

Since I am unable to install the software at the moment, I cannot complete my assessment of its functionality. Once we get that squared away, I'll be able to finish up my review :)
