### 2021-08-25 #3635 @crvernon
@crhea93 I believe you accidentally linked to issue number 11 in the JOSS repository registry instead of one of the issues that is currently in the author's repository here:  https://github.com/grant-m-s/AstronomicAL/issues

Would you mind  editing your comment to reference the issue in the author's repo that you are referring to?  Thanks!
