### 2021-08-25 grant-m-s/AstronomicAL#9 @rmorgan10
I think the length is a-ok, no need to trim further. I'm picturing an additional paragraph about the state of the field that concludes with something along the lines of "... and astronomicAL's goal is to satisfy these needs of the field." Then go into your current statement of need text which details the main functionalities. And I don't think one additional paragraph would hurt the overall length.
