### 2021-08-25 grant-m-s/AstronomicAL#11 @rmorgan10
Yup that is perfect. You can close this issue :)
