### 2021-08-25 grant-m-s/AstronomicAL#12 @rmorgan10
When I run from the top-level directory, I get 12 failures:

```
platform darwin -- Python 3.9.6, pytest-6.2.4, py-1.10.0, pluggy-0.13.1
rootdir: /Users/rmorgan/garb/astronomicAL
collected 122 items

tests/all_test.py ...................................................... [ 44%]
..........FFFFFF.F.F........F..FFF..................................                                                                                                  [100%]

================================================================================= FAILURES ==================================================================================
_____________________________________________________________ TestDashboards.test_selected_source_init_selected _____________________________________________________________

self = <all_test.TestDashboards object at 0x143120760>

    def test_selected_source_init_selected(self):

        data = self._create_test_df()
        config.main_df = data
        data = data.iloc[5]
        src = ColumnDataSource({str(c): [v] for c, v in data.items()})
>       selected_source = SelectedSourceDashboard(src=src, close_button=None)

tests/all_test.py:1275:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
astronomicAL/dashboard/selected_source.py:83: in __init__
    self._add_selected_info()
astronomicAL/dashboard/selected_source.py:153: in _add_selected_info
    self.panel()
astronomicAL/dashboard/selected_source.py:455: in panel
    Process(target=self._update_default_images).start()
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/process.py:121: in start
    self._popen = self._Popen(self)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:224: in _Popen
    return _default_context.get_context().Process._Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:284: in _Popen
    return Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:32: in __init__
    super().__init__(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_fork.py:19: in __init__
    self._launch(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:47: in _launch
    reduction.dump(process_obj, fp)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

obj = <Process name='Process-1' parent=53369 initial>, file = <_io.BytesIO object at 0x142feff40>, protocol = None

    def dump(obj, file, protocol=None):
        '''Replacement for pickle.dump() using ForkingPickler.'''
>       ForkingPickler(file, protocol).dump(obj)
E       _pickle.PicklingError: Can't pickle <function ColumnDataSource.<lambda> at 0x10dcd5a60>: attribute lookup ColumnDataSource.<lambda> on bokeh.models.sources failed

../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/reduction.py:60: PicklingError
__________________________________________________ TestDashboards.test_selected_source_check_history_from_none_to_selected __________________________________________________

self = <all_test.TestDashboards object at 0x143314640>

    def test_selected_source_check_history_from_none_to_selected(self):
        data = self._create_test_df()
        config.main_df = data
        src = ColumnDataSource()
        selected_source = SelectedSourceDashboard(src=src, close_button=None)

        assert selected_source.selected_history == []

        data_selected = data.iloc[72]
>       src.data = {str(c): [v] for c, v in data_selected.items()}

tests/all_test.py:1311:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
../../anaconda3/envs/astronomical/lib/python3.9/site-packages/bokeh/core/has_props.py:278: in __setattr__
    super().__setattr__(name, value)
../../anaconda3/envs/astronomical/lib/python3.9/site-packages/bokeh/core/property/descriptors.py:978: in __set__
    self._internal_set(obj, value, hint=hint, setter=setter)
../../anaconda3/envs/astronomical/lib/python3.9/site-packages/bokeh/core/property/descriptors.py:763: in _internal_set
    self._real_set(obj, old, value, hint=hint, setter=setter)
../../anaconda3/envs/astronomical/lib/python3.9/site-packages/bokeh/core/property/descriptors.py:832: in _real_set
    self._trigger(obj, old, value, hint=hint, setter=setter)
../../anaconda3/envs/astronomical/lib/python3.9/site-packages/bokeh/core/property/descriptors.py:909: in _trigger
    obj.trigger(self.name, old, value, hint, setter)
../../anaconda3/envs/astronomical/lib/python3.9/site-packages/bokeh/model.py:664: in trigger
    super().trigger(attr, old, new, hint=hint, setter=setter)
../../anaconda3/envs/astronomical/lib/python3.9/site-packages/bokeh/util/callback_manager.py:159: in trigger
    invoke()
../../anaconda3/envs/astronomical/lib/python3.9/site-packages/bokeh/util/callback_manager.py:155: in invoke
    callback(attr, old, new)
astronomicAL/dashboard/selected_source.py:182: in _panel_cb
    self.panel()
astronomicAL/dashboard/selected_source.py:455: in panel
    Process(target=self._update_default_images).start()
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/process.py:121: in start
    self._popen = self._Popen(self)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:224: in _Popen
    return _default_context.get_context().Process._Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:284: in _Popen
    return Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:32: in __init__
    super().__init__(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_fork.py:19: in __init__
    self._launch(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:47: in _launch
    reduction.dump(process_obj, fp)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

obj = <Process name='Process-2' parent=53369 initial>, file = <_io.BytesIO object at 0x1431b0400>, protocol = None

    def dump(obj, file, protocol=None):
        '''Replacement for pickle.dump() using ForkingPickler.'''
>       ForkingPickler(file, protocol).dump(obj)
E       _pickle.PicklingError: Can't pickle <function ColumnDataSource.<lambda> at 0x10dcd5a60>: attribute lookup ColumnDataSource.<lambda> on bokeh.models.sources failed

../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/reduction.py:60: PicklingError
____________________________________________ TestDashboards.test_selected_source_check_history_from_selected_to_selected_unique _____________________________________________

self = <all_test.TestDashboards object at 0x143551be0>

    def test_selected_source_check_history_from_selected_to_selected_unique(self):
        data = self._create_test_df()
        data_selected = data.iloc[72]
        src = ColumnDataSource({str(c): [v] for c, v in data_selected.items()})
>       selected_source = SelectedSourceDashboard(src=src, close_button=None)

tests/all_test.py:1319:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
astronomicAL/dashboard/selected_source.py:83: in __init__
    self._add_selected_info()
astronomicAL/dashboard/selected_source.py:153: in _add_selected_info
    self.panel()
astronomicAL/dashboard/selected_source.py:455: in panel
    Process(target=self._update_default_images).start()
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/process.py:121: in start
    self._popen = self._Popen(self)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:224: in _Popen
    return _default_context.get_context().Process._Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:284: in _Popen
    return Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:32: in __init__
    super().__init__(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_fork.py:19: in __init__
    self._launch(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:47: in _launch
    reduction.dump(process_obj, fp)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

obj = <Process name='Process-3' parent=53369 initial>, file = <_io.BytesIO object at 0x143b699f0>, protocol = None

    def dump(obj, file, protocol=None):
        '''Replacement for pickle.dump() using ForkingPickler.'''
>       ForkingPickler(file, protocol).dump(obj)
E       _pickle.PicklingError: Can't pickle <function ColumnDataSource.<lambda> at 0x10dcd5a60>: attribute lookup ColumnDataSource.<lambda> on bokeh.models.sources failed

../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/reduction.py:60: PicklingError
_________________________________________ TestDashboards.test_selected_source_check_history_from_selected_to_selected_same_id_head __________________________________________

self = <all_test.TestDashboards object at 0x1439bdd30>

    def test_selected_source_check_history_from_selected_to_selected_same_id_head(self):
        data = self._create_test_df()
        config.main_df = data
        data_selected = data.iloc[72]
        src = ColumnDataSource({str(c): [v] for c, v in data_selected.items()})
>       selected_source = SelectedSourceDashboard(src=src, close_button=None)

tests/all_test.py:1333:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
astronomicAL/dashboard/selected_source.py:83: in __init__
    self._add_selected_info()
astronomicAL/dashboard/selected_source.py:153: in _add_selected_info
    self.panel()
astronomicAL/dashboard/selected_source.py:455: in panel
    Process(target=self._update_default_images).start()
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/process.py:121: in start
    self._popen = self._Popen(self)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:224: in _Popen
    return _default_context.get_context().Process._Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:284: in _Popen
    return Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:32: in __init__
    super().__init__(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_fork.py:19: in __init__
    self._launch(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:47: in _launch
    reduction.dump(process_obj, fp)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

obj = <Process name='Process-4' parent=53369 initial>, file = <_io.BytesIO object at 0x1431e6cc0>, protocol = None

    def dump(obj, file, protocol=None):
        '''Replacement for pickle.dump() using ForkingPickler.'''
>       ForkingPickler(file, protocol).dump(obj)
E       _pickle.PicklingError: Can't pickle <function ColumnDataSource.<lambda> at 0x10dcd5a60>: attribute lookup ColumnDataSource.<lambda> on bokeh.models.sources failed

../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/reduction.py:60: PicklingError
_____________________________________________ TestDashboards.test_selected_source_check_history_from_selected_to_selected_no_id _____________________________________________

self = <all_test.TestDashboards object at 0x14325bb80>

    def test_selected_source_check_history_from_selected_to_selected_no_id(self):
        data = self._create_test_df()
        config.main_df = data
        data_selected = data.iloc[72]
        src = ColumnDataSource({str(c): [v] for c, v in data_selected.items()})
>       selected_source = SelectedSourceDashboard(src=src, close_button=None)

tests/all_test.py:1347:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
astronomicAL/dashboard/selected_source.py:83: in __init__
    self._add_selected_info()
astronomicAL/dashboard/selected_source.py:153: in _add_selected_info
    self.panel()
astronomicAL/dashboard/selected_source.py:455: in panel
    Process(target=self._update_default_images).start()
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/process.py:121: in start
    self._popen = self._Popen(self)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:224: in _Popen
    return _default_context.get_context().Process._Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:284: in _Popen
    return Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:32: in __init__
    super().__init__(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_fork.py:19: in __init__
    self._launch(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:47: in _launch
    reduction.dump(process_obj, fp)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

obj = <Process name='Process-5' parent=53369 initial>, file = <_io.BytesIO object at 0x142f81590>, protocol = None

    def dump(obj, file, protocol=None):
        '''Replacement for pickle.dump() using ForkingPickler.'''
>       ForkingPickler(file, protocol).dump(obj)
E       _pickle.PicklingError: Can't pickle <function ColumnDataSource.<lambda> at 0x10dcd5a60>: attribute lookup ColumnDataSource.<lambda> on bokeh.models.sources failed

../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/reduction.py:60: PicklingError
______________________________________ TestDashboards.test_selected_source_check_history_from_selected_to_selected_same_id_throughout _______________________________________

self = <all_test.TestDashboards object at 0x14398bc40>

    def test_selected_source_check_history_from_selected_to_selected_same_id_throughout(
        self,
    ):
        data = self._create_test_df()
        config.main_df = data
        data_selected = data.iloc[72]
        src = ColumnDataSource({str(c): [v] for c, v in data_selected.items()})
>       selected_source = SelectedSourceDashboard(src=src, close_button=None)

tests/all_test.py:1364:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
astronomicAL/dashboard/selected_source.py:83: in __init__
    self._add_selected_info()
astronomicAL/dashboard/selected_source.py:153: in _add_selected_info
    self.panel()
astronomicAL/dashboard/selected_source.py:455: in panel
    Process(target=self._update_default_images).start()
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/process.py:121: in start
    self._popen = self._Popen(self)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:224: in _Popen
    return _default_context.get_context().Process._Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:284: in _Popen
    return Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:32: in __init__
    super().__init__(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_fork.py:19: in __init__
    self._launch(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:47: in _launch
    reduction.dump(process_obj, fp)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

obj = <Process name='Process-6' parent=53369 initial>, file = <_io.BytesIO object at 0x142f81950>, protocol = None

    def dump(obj, file, protocol=None):
        '''Replacement for pickle.dump() using ForkingPickler.'''
>       ForkingPickler(file, protocol).dump(obj)
E       _pickle.PicklingError: Can't pickle <function ColumnDataSource.<lambda> at 0x10dcd5a60>: attribute lookup ColumnDataSource.<lambda> on bokeh.models.sources failed

../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/reduction.py:60: PicklingError
_____________________________________________________ TestDashboards.test_selected_source_empty_selected_from_selected ______________________________________________________

self = <all_test.TestDashboards object at 0x143b2a460>

    def test_selected_source_empty_selected_from_selected(self):
        data = self._create_test_df()
        config.main_df = data
        data_selected = data.iloc[72]
        src = ColumnDataSource({str(c): [v] for c, v in data_selected.items()})
>       selected_source = SelectedSourceDashboard(src=src, close_button=None)

tests/all_test.py:1394:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
astronomicAL/dashboard/selected_source.py:83: in __init__
    self._add_selected_info()
astronomicAL/dashboard/selected_source.py:153: in _add_selected_info
    self.panel()
astronomicAL/dashboard/selected_source.py:455: in panel
    Process(target=self._update_default_images).start()
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/process.py:121: in start
    self._popen = self._Popen(self)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:224: in _Popen
    return _default_context.get_context().Process._Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/context.py:284: in _Popen
    return Popen(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:32: in __init__
    super().__init__(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_fork.py:19: in __init__
    self._launch(process_obj)
../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/popen_spawn_posix.py:47: in _launch
    reduction.dump(process_obj, fp)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

obj = <Process name='Process-7' parent=53369 initial>, file = <_io.BytesIO object at 0x14321d2c0>, protocol = None

    def dump(obj, file, protocol=None):
        '''Replacement for pickle.dump() using ForkingPickler.'''
>       ForkingPickler(file, protocol).dump(obj)
E       _pickle.PicklingError: Can't pickle <function ColumnDataSource.<lambda> at 0x10dcd5a60>: attribute lookup ColumnDataSource.<lambda> on bokeh.models.sources failed

../../anaconda3/envs/astronomical/lib/python3.9/multiprocessing/reduction.py:60: PicklingError
_______________________________________________ TestDashboards.test_selected_source_check_valid_select_from_selected_is_valid _______________________________________________

[…]
