### 2021-08-25 grant-m-s/AstronomicAL#12 @rmorgan10
Seems like all the remaining issues are pickle related. Maybe this hints at some sort of platform dependence in the pickling?
