### 2021-08-25 grant-m-s/AstronomicAL#9 @rmorgan10
By the way one additional thing (though completely optional) is to use a colorblind friendly colormap in Figure 3. I put a red-green colorblind filter on my screen and couldn't tell the difference between the red and green regions, so the gradient of the colormap was lost.
