### 2021-08-25 0a71ea5
Updated README to add line to conda install
Fix for #10 .
Added new `conda` line to Readme to allow for all libraries to be found on install.
