### 2021-08-25 9a66cff
Updated Docs page to add line to conda install
Fix for #10 .
Added new `conda` line to index.rst to allow for all libraries to be found on install.
