### 2021-08-28 aa7144b
Updates to paper in response to review issue #9 .
