### 2021-08-28 ada3318
Updates to paper in response to review issue #9 .
