+++ paper/paper.bib
@@ -264,3 +264,51 @@ author={
     Xiaoqiang~Zheng},
   year={2015},
 }
+
+@inproceedings{baldridge2009well,
+  title={How well does active learning actually work? Time-based evaluation of cost-reduction strategies for language documentation.},
+  author={Baldridge, Jason and Palmer, Alexis},
+  booktitle={Proceedings of the 2009 Conference on Empirical Methods in Natural Language Processing},
+  pages={296--305},
+  year={2009}
+}
+
+@inproceedings{settles2008analysis,
+  title={An analysis of active learning strategies for sequence labeling tasks},
+  author={Settles, Burr and Craven, Mark},
+  booktitle={Proceedings of the 2008 Conference on Empirical Methods in Natural Language Processing},
+  pages={1070--1079},
+  year={2008}
+}
+
+@inproceedings{gal2017deep,
+  title={Deep bayesian active learning with image data},
+  author={Gal, Yarin and Islam, Riashat and Ghahramani, Zoubin},
+  booktitle={International Conference on Machine Learning},
+  pages={1183--1192},
+  year={2017},
+  organization={PMLR}
+}
+
+
+@article{walmsley2020galaxy,
+  title={Galaxy Zoo: probabilistic morphology through Bayesian CNNs and active learning},
+  author={Walmsley, Mike and Smith, Lewis and Lintott, Chris and Gal, Yarin and Bamford, Steven and Dickinson, Hugh and Fortson, Lucy and Kruk, Sandor and Masters, Karen and Scarlata, Claudia and others},
+  journal={Monthly Notices of the Royal Astronomical Society},
+  volume={491},
+  number={2},
+  pages={1554--1574},
+  year={2020},
+  publisher={Oxford University Press}
+}
+
+@article{eisenstein2020active,
+  title={Active machine learning helps drug hunters tackle biology.},
+  author={Eisenstein, Michael},
+  journal={Nature biotechnology},
+  volume={38},
+  number={5},
+  pages={512--515},
+  year={2020},
+  publisher={Nature Publishing Group}
+}
