### 2021-08-28 cd4be9d
Updates to paper in response to review issue #9 .
