### 2021-08-28 grant-m-s/AstronomicAL#9 @grant-m-s
I've pushed an updated paper (viewable on the main review issue). Let me know if that is what you had in mind.
