### 2021-08-30 grant-m-s/AstronomicAL#9 @rmorgan10
Your changes are exactly what I had in mind. Looks good to me! Go ahead and close this issue and we can reopen it if any of the other reviewers have paper comments.
