### 2021-08-30 grant-m-s/AstronomicAL#12 @rmorgan10
Gotcha. Yeah I tried two things:

1. Python 3.7 + not adding the lines above
1. Python 3.9 + adding the lines above

Both resulted in all tests passing, and in the second case I see the reported crashing of processes. Personally, I think the safest thing to do is to add python==3.7 to the requirements.txt file, but as long as your tests pass I'm not too worried about crashing processes that only occur when tests are run.

So two options:
1. Add python==3.7 to requirements.txt, or
1. Add something like the following to the top of your test modules:

```
if sys.version_info >= (3, 8):
    # Account for python version dependence in the default multiprocessing
    # starting of processes: https://github.com/pytest-dev/pytest-flask/issues/104
    import multiprocessing
    multiprocessing.set_start_method("fork")
```
