### 2021-08-30 grant-m-s/AstronomicAL#10 @rmorgan10
One more suggestion for the installation: Can you please add pytest to the requirements.txt file? Currently I have to pip install into the astronomical environment to run the tests
