### 2021-08-30 b5a6a65
Added pytest to req file to address issue #10
