### 2021-08-30 3a20a2f
Added pytest to req file to address issue #10
