+++ requirements.txt
@@ -6,3 +6,4 @@ modAL>=0.4.1
 holoviews==1.14.1
 datashader==0.12.0
 coverage>=5.5
+pytest>=6.2.4
