### 2021-08-30 aebd112
Added pytest to req file to address issue #10
