### 2021-08-30 8c33b46
Account for multiprocessing change in 3.8+
Fix for issue #12 .
