### 2021-08-30 7f057b5
Account for multiprocessing change in 3.8+
Fix for issue #12 .
