### 2021-08-30 6789c4b
Account for multiprocessing change in 3.8+
Fix for issue #12 .
