### 2021-08-30 grant-m-s/AstronomicAL#12 @grant-m-s
Awesome, glad its now working. I decided on going for option 2.
Are you happy for this now to be closed?
