### 2021-09-02 grant-m-s/AstronomicAL#10 @grant-m-s
Awesome! Thanks for this.

I think this is now closable.
