### 2021-09-02 #3635 @crvernon
@grant-m-s - I'll take a look at the paper document later today and build a list of any edits I think need to be made (if there are any).

Once I deem everything is good to go, I'll get you to generate a release and an accompanying DOI as linked with a service like Zenodo that corresponds to the final review version.

Lastly, once all has been accounted for, I'll make my recommendation to the editor-in-chief and they will conduct a final pass and move forward with full acceptance and publication if they are in agreement with my recommendation.
