### 2021-09-02 #3635 @crvernon
@grant-m-s your `joss-paper-branch` is 9 commits behind and 5 commits ahead of the `master` branch.  Some of your changes per review comments were only made to the `master` and the paper directory resides on the `joss-paper-branch`.  Are you planning on merging `joss-paper-branch` into `master`?
