### 2021-09-02 #3635 @crvernon
@grant-m-s - In your [Citing the Software](https://astronomical.readthedocs.io/en/latest/content/other/citing.html#id9) documentation, you mention that:

> astronomicAL has also been published as an arXiv_ preprint.

Could you please give me a link to that preprint?
