### 2021-09-02 #3635 @grant-m-s
@crvernon That was added at the beginning of the docs being created as we assumed we would upload but we never actually did. Would you rather that I remove that?
