### 2021-09-02 #3635 @crvernon
Yes, if the `arXiv` submission was never made it would be good to remove this to support the originality of this work.  Thank you!
