### 2021-09-02 f34b971
Remove placeholder for Arxiv citation
