### 2021-09-02 #3635 @crvernon
@grant-m-s - Your paper is looking great!  Here are several things I would like you to address from my review of your paper before we move on:

- [ ] Please check the missing DOI in the references (see the above `whedon check references` output ☝️)
- [ ] Please review any other references to ensure that you have added in their DOI to the citation if there is one available.  Sometimes these get excluded from citation exports.
- [ ] Figure 1 is placed in the document before it is referred to in the text
- [ ] Figures 1 and 3 are never referenced in the text
- [ ] Please take one last look at your author names and their affiliations to make sure all is well before we move forward
