+++ paper/paper.bib
@@ -152,7 +152,8 @@ url = {https://bokeh.pydata.org/en/latest/},
   number={1},
   pages={1--114},
   year={2012},
-  publisher={Morgan \& Claypool Publishers}
+  publisher={Morgan \& Claypool Publishers},
+  doi={10.2200/S00429ED1V01Y201207AIM018}
 }
 
 @ARTICLE{Fotopoulou2018CPz,
@@ -263,6 +264,7 @@ author={
     Yuan~Yu and
     Xiaoqiang~Zheng},
   year={2015},
+  doi={10.5281/zenodo.4724125}
 }
 
 @inproceedings{baldridge2009well,
@@ -270,7 +272,8 @@ author={
   author={Baldridge, Jason and Palmer, Alexis},
   booktitle={Proceedings of the 2009 Conference on Empirical Methods in Natural Language Processing},
   pages={296--305},
-  year={2009}
+  year={2009},
+  doi={10.5555/1699510.1699549}
 }
 
 @inproceedings{settles2008analysis,
@@ -278,7 +281,8 @@ author={
   author={Settles, Burr and Craven, Mark},
   booktitle={Proceedings of the 2008 Conference on Empirical Methods in Natural Language Processing},
   pages={1070--1079},
-  year={2008}
+  year={2008},
+  doi={10.3115/1613715.1613855}
 }
 
 @inproceedings{gal2017deep,
@@ -287,7 +291,8 @@ author={
   booktitle={International Conference on Machine Learning},
   pages={1183--1192},
   year={2017},
-  organization={PMLR}
+  organization={PMLR},
+  doi={10.5555/3305381.3305504}
 }
 
 
@@ -299,7 +304,8 @@ author={
   number={2},
   pages={1554--1574},
   year={2020},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/mnras/stz2816}
 }
 
 @article{eisenstein2020active,
@@ -310,5 +316,6 @@ author={
   number={5},
   pages={512--515},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/s41587-020-0521-4}
 }
