### 2021-09-02 #3635 @grant-m-s
@crvernon so I've added all doi's I can find, its mainly the software packages that dont have a zenodo that I'm missing.

I've referenced all the figures in the text now.

Figure 1 is still ahead in the text as I believe it fits much better there rather than half way through the section. Let me know if that's against any style requirements.

Authors and Affiliations are all good.
