### 2021-09-02 #3635 @crvernon
@grant-m-s - I think the figure referencing looks good.
