### 2021-09-02 #3635 @grant-m-s
hmm both seem to be ACM doi's that aren't matched anywhere else besides the ACM website.

should I remove them?
