### 2021-09-02 #3635 @crvernon
Yes, please remove them.  It looks like those repos only site the tag on GitHub and not a DOI.  Please pull the two invalid ones and we will proceed.
