+++ paper/paper.bib
@@ -272,8 +272,7 @@ author={
   author={Baldridge, Jason and Palmer, Alexis},
   booktitle={Proceedings of the 2009 Conference on Empirical Methods in Natural Language Processing},
   pages={296--305},
-  year={2009},
-  doi={10.5555/1699510.1699549}
+  year={2009}
 }
 
 @inproceedings{settles2008analysis,
@@ -291,8 +290,7 @@ author={
   booktitle={International Conference on Machine Learning},
   pages={1183--1192},
   year={2017},
-  organization={PMLR},
-  doi={10.5555/3305381.3305504}
+  organization={PMLR}
 }
 
 
