### 2021-09-09 #3635 @grant-m-s
@crvernon If I were to also upload this to Arxiv, am I able to upload the paper with JOSS branding? (For example the main paper pdf or the local docker created version). Thanks!
