### 2021-09-09 #3635 @crvernon
@grant-m-s I think this is more of a question for the arXiv staff to determine permissiveness.  Your JOSS paper is now licensed as follows (and as listed on your publication):

> Authors of papers retain copyright and release the work under a Creative Commons Attribution 4.0 International License ([CC BY 4.0](https://creativecommons.org/licenses/by/4.0/)).

Have a great day!
